Cleanup Fixed the compiler form being unrecognized by VS
